More `wast` refactorings in preparation for components
A few small changes are included here as refactorings in the `wast`
crate which are either fixing bugs or making room for the component
implementation:

* The `Encode` trait and common base impls are moved to the crate root
  for component encoding to use the same trait as well.

* The `Peek for InlineImport` type for core wasm was updated to require
  two strings instead of optionally having only one string. This is a
  bug fix relative to #576

* The `resolve` method for core wasm modules has been changed to take
  `&mut Vec<ModuleField>` as input. The distinction of text/binary is
  now made at the `Module::resolve` and `Module::encode` layer.

* The `Namespace` type has been extracted to a top-level, but still
  crate-private, module. This will be used for name resolution with
  components as well.

* A `Wat::encode` method was added since it won't be as simple as
  `foo.module.encode()` in the future.
