+++ crates/wast/src/core/binary.rs
@@ -1,14 +1,8 @@
 use crate::core::*;
+use crate::encode::Encode;
 use crate::token::*;
 
-pub fn encode(module: &Module<'_>) -> Vec<u8> {
-    match &module.kind {
-        ModuleKind::Text(fields) => encode_fields(&module.id, &module.name, fields),
-        ModuleKind::Binary(bytes) => bytes.iter().flat_map(|b| b.iter().cloned()).collect(),
-    }
-}
-
-fn encode_fields(
+pub fn encode(
     module_id: &Option<Id<'_>>,
     module_name: &Option<NameAnnotation<'_>>,
     fields: &[ModuleField<'_>],
@@ -131,75 +125,6 @@ impl Encoder<'_> {
     }
 }
 
-pub(crate) trait Encode {
-    fn encode(&self, e: &mut Vec<u8>);
-}
-
-impl<T: Encode + ?Sized> Encode for &'_ T {
-    fn encode(&self, e: &mut Vec<u8>) {
-        T::encode(self, e)
-    }
-}
-
-impl<T: Encode> Encode for [T] {
-    fn encode(&self, e: &mut Vec<u8>) {
-        self.len().encode(e);
-        for item in self {
-            item.encode(e);
-        }
-    }
-}
-
-impl<T: Encode> Encode for Vec<T> {
-    fn encode(&self, e: &mut Vec<u8>) {
-        <[T]>::encode(self, e)
-    }
-}
-
-impl Encode for str {
-    fn encode(&self, e: &mut Vec<u8>) {
-        self.len().encode(e);
-        e.extend_from_slice(self.as_bytes());
-    }
-}
-
-impl Encode for usize {
-    fn encode(&self, e: &mut Vec<u8>) {
-        assert!(*self <= u32::max_value() as usize);
-        (*self as u32).encode(e)
-    }
-}
-
-impl Encode for u8 {
-    fn encode(&self, e: &mut Vec<u8>) {
-        e.push(*self);
-    }
-}
-
-impl Encode for u32 {
-    fn encode(&self, e: &mut Vec<u8>) {
-        leb128::write::unsigned(e, (*self).into()).unwrap();
-    }
-}
-
-impl Encode for i32 {
-    fn encode(&self, e: &mut Vec<u8>) {
-        leb128::write::signed(e, (*self).into()).unwrap();
-    }
-}
-
-impl Encode for u64 {
-    fn encode(&self, e: &mut Vec<u8>) {
-        leb128::write::unsigned(e, (*self).into()).unwrap();
-    }
-}
-
-impl Encode for i64 {
-    fn encode(&self, e: &mut Vec<u8>) {
-        leb128::write::signed(e, *self).unwrap();
-    }
-}
-
 impl Encode for FunctionType<'_> {
     fn encode(&self, e: &mut Vec<u8>) {
         self.params.len().encode(e);
@@ -256,13 +181,6 @@ impl Encode for Type<'_> {
 impl Encode for Option<Id<'_>> {
     fn encode(&self, _e: &mut Vec<u8>) {
         // used for parameters in the tuple impl as well as instruction labels
-    }
-}
-
-impl<T: Encode, U: Encode> Encode for (T, U) {
-    fn encode(&self, e: &mut Vec<u8>) {
-        self.0.encode(e);
-        self.1.encode(e);
     }
 }
